Minor performance improvements.
Changing min dist for collision detection from 1 to .1.
Moving the if check out of the function so that it doesn't leave
function scope unless it has to (this added about 1 fps in my quick
testing).
Saw an 18 fps versus 15 fps with this compared to master for one run.
Splitting a physics-type function out of viewport and into physics.