Implement Vector3d as TypedArray Views on individual buffers
Everywhere that vectors are used for physics, (position, velocity, and
acceleration) appear as arrays of length 3.  The Vector library is not
object oriented, but executes procedures on existing arrays.

This implementation was chosen to be forwards compatible with a shared
buffer implementation, where all the time-critical physics vectors will
be stored.

Also, more camelCase, fewer snake_cases.