Introduce orbital inclination
Inclination can now be set via the particle configuration.  An
inclination is described by the "axis of nodes" (a Vector3d)
and the inclination itself, measured from the xy-plane.

Calculations done by new Vector3d.rotateAboutAxis
Tweaked initial positions in particle.configure so that the
center of mass of the system is near [0,0,0].