Add vector math tests, standardize function names
Performs tests on randomly generated vectors for correctness of known
relations. Mutators are renamed to contain "Me", ie "rotateMe",
"incrementMe", etc. Functions are reordered such that "setters" are
adjacent to "getters".  Vector3d.js and its tests are linted.

API:

Added a short block for aliasing the functions with the old names.
Added compatibility function wrapper for rotateAbout (just to keep
things running and testable).  This calls rotateMe, which takes an
object with rotation details: pivotPoint (location in space about
which to rotate), axis: the axis of rotation (it can be thought of as
passing through the pivot point), and the angle of rotation.

Testing:

Extends Math to include an approximate equals, Math.isClose(a,b).

Vector3d tests are found in tests/Vector3dtests.js.

tests/Vector3dtests.html runs the tests.

The first tests are testAxes, testing if axes (unit vectors (1,0,0),
(0,1,0), (0,0,1)) are perpendicular, length one, etc.

Tested vector adding (subtracting) for commutativity
(anti)-commutativity.