Add special test cases, including the zero vector
Coordinate axis vectors are added to both the unit vector list and
general vector list.  The zero vector is also added to the general list.
This is a bit of a torture test for the library, since both numerical
stability of perpendicularity and parallel(... ness?) are very
sensitive to the choice of how the lengths, angles, etc are evaluated.

The test also forces a decision as to whether the zero vector should
be perpendicular to every vector (good reason to say yes). Whether
the zero vector should be coaxial to every vector, I'm not sure, but
I picked yes.

(Vectors are coaxial if they point in either the same or exactly
opposite directions.  "Parallel" is reserved for when two coaxial
vectors point in the same direction.)

Also, replaced the test of opening angle (which is complicated,
expensive, and wrong for the zero vector) with a test for invariance of
the dot product.